refactor user services
